+++ druid/src/app.rs
@@ -20,9 +20,7 @@ use crate::shell::{Application, Error as PlatformError, WindowBuilder, WindowHan
 use crate::widget::LabelText;
 use crate::win_handler::{AppHandler, AppState};
 use crate::window::WindowId;
-use crate::{
-    theme, AppDelegate, Data, DruidHandler, Env, LocalizedString, MenuDesc, Widget, WidgetExt,
-};
+use crate::{AppDelegate, Data, DruidHandler, Env, LocalizedString, MenuDesc, Widget, WidgetExt};
 
 use druid_shell::WindowState;
 
@@ -111,14 +109,6 @@ impl<T: Data> AppLauncher<T> {
         self.ext_event_host.make_sink()
     }
 
-    /// Paint colorful rectangles for layout debugging.
-    ///
-    /// The rectangles are drawn around each widget's layout rect.
-    #[deprecated(since = "0.5.0", note = "Use WidgetExt::debug_paint_layout instead.")]
-    pub fn debug_paint_layout(self) -> Self {
-        self
-    }
-
     /// Build the windows and start the runloop.
     ///
     /// Returns an error if a window cannot be instantiated. This is usually
@@ -126,7 +116,7 @@ impl<T: Data> AppLauncher<T> {
     pub fn launch(mut self, data: T) -> Result<(), PlatformError> {
         let app = Application::new()?;
 
-        let mut env = theme::init();
+        let mut env = Env::default();
         if let Some(f) = self.env_setup.take() {
             f(&mut env, &data);
         }