+++ druid/src/theme.rs
@@ -15,6 +15,7 @@
 //! Theme keys and initial values.
 
 #![allow(missing_docs)]
+
 use crate::piet::Color;
 
 use crate::{Env, FontDescriptor, FontFamily, Key};
@@ -60,6 +61,17 @@ pub const BORDERED_WIDGET_HEIGHT: Key<f64> =
 
 pub const TEXTBOX_BORDER_RADIUS: Key<f64> = Key::new("org.linebender.druid.theme.textbox_radius");
 
+/// The default horizontal spacing between widgets.
+pub const WIDGET_PADDING_HORIZONTAL: Key<f64> =
+    Key::new("org.linebender.druid.theme.widget-padding-h");
+/// The default vertical spacing between widgets.
+pub const WIDGET_PADDING_VERTICAL: Key<f64> =
+    Key::new("org.linebender.druid.theme.widget-padding-v");
+/// The default internal (horizontal) padding for visually distinct components
+/// of a widget; for instance between a checkbox and its label.
+pub const WIDGET_CONTROL_COMPONENT_PADDING: Key<f64> =
+    Key::new("org.linebender.druid.theme.widget-padding-control-label");
+
 pub const SCROLLBAR_COLOR: Key<Color> = Key::new("org.linebender.druid.theme.scrollbar_color");
 pub const SCROLLBAR_BORDER_COLOR: Key<Color> =
     Key::new("org.linebender.druid.theme.scrollbar_border_color");
@@ -74,9 +86,8 @@ pub const SCROLLBAR_EDGE_WIDTH: Key<f64> =
     Key::new("org.linebender.druid.theme.scrollbar_edge_width");
 
 /// An initial theme.
-pub fn init() -> Env {
-    Env::default()
-        .adding(WINDOW_BACKGROUND_COLOR, Color::rgb8(0x29, 0x29, 0x29))
+pub(crate) fn add_to_env(env: Env) -> Env {
+    env.adding(WINDOW_BACKGROUND_COLOR, Color::rgb8(0x29, 0x29, 0x29))
         .adding(LABEL_COLOR, Color::rgb8(0xf0, 0xf0, 0xea))
         .adding(PLACEHOLDER_COLOR, Color::rgb8(0x80, 0x80, 0x80))
         .adding(PRIMARY_LIGHT, Color::rgb8(0x5c, 0xc4, 0xff))
@@ -109,8 +120,16 @@ pub fn init() -> Env {
         .adding(SCROLLBAR_PAD, 2.)
         .adding(SCROLLBAR_RADIUS, 5.)
         .adding(SCROLLBAR_EDGE_WIDTH, 1.)
+        .adding(WIDGET_PADDING_VERTICAL, 10.0)
+        .adding(WIDGET_PADDING_HORIZONTAL, 8.0)
+        .adding(WIDGET_CONTROL_COMPONENT_PADDING, 4.0)
         .adding(
             UI_FONT,
             FontDescriptor::new(FontFamily::SYSTEM_UI).with_size(15.0),
         )
+}
+
+#[deprecated(since = "0.7.0", note = "use Env::default() instead")]
+pub fn init() -> Env {
+    Env::default()
 }